% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_frechet_mean.R
\name{pfm_objective_value}
\alias{pfm_objective_value}
\title{Evaluate the Power Frechet Objective at a Given Point}
\usage{
pfm_objective_value(x, X, a = 2)
}
\arguments{
\item{x}{Numeric vector of length \eqn{d}.}

\item{X}{Numeric \eqn{n \times d} matrix of data points.}

\item{a}{Power parameter \eqn{a > 0} (default \code{2}).}
}
\value{
Scalar numeric objective value.
}
\description{
Evaluates \eqn{f(x) = \sum_{i=1}^{n} \|x_i - x\|^a} at an arbitrary
point \code{x}.
}
\examples{
X <- matrix(1:6, nrow = 3)
pfm_objective_value(c(2, 3), X, a = 2)

}
