% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_frechet_mean.R
\name{print.power_frechet_mean}
\alias{print.power_frechet_mean}
\title{Print method for power_frechet_mean objects}
\usage{
\method{print}{power_frechet_mean}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"power_frechet_mean"}.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for power_frechet_mean objects
}
