% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_huber_frechet_mean.R
\name{pseudo_huber_objective_value}
\alias{pseudo_huber_objective_value}
\title{Evaluate the Pseudo-Huber Frechet Objective at a Given Point}
\usage{
pseudo_huber_objective_value(x, X, delta)
}
\arguments{
\item{x}{Numeric vector of length \eqn{d}.}

\item{X}{Numeric \eqn{n \times d} matrix of data points.}

\item{delta}{Pseudo-Huber threshold \eqn{\delta > 0}.}
}
\value{
Scalar numeric objective value.
}
\description{
Evaluates \eqn{f(x) = \sum_{i=1}^{n} \rho_\delta(\|x_i - x\|)} at an
arbitrary point \code{x}, where \eqn{\rho_\delta(r) = \delta^2(\sqrt{1 +
(r/\delta)^2} - 1)} is the pseudo-Huber loss.
}
\examples{
X <- matrix(1:6, nrow = 3)
pseudo_huber_objective_value(c(2, 3), X, delta = 1)

}
\seealso{
\code{\link[=pseudo_huber_frechet_mean]{pseudo_huber_frechet_mean()}}
}
